+++ internal/cluster/clusterpb/ObjectMeta.go
@@ -270,8 +270,20 @@ func (rcv *ObjectMeta) TagsLength() int {
 	return 0
 }
 
+func (rcv *ObjectMeta) StripeBytes() uint32 {
+	o := flatbuffers.UOffsetT(rcv._tab.Offset(38))
+	if o != 0 {
+		return rcv._tab.GetUint32(o + rcv._tab.Pos)
+	}
+	return 0
+}
+
+func (rcv *ObjectMeta) MutateStripeBytes(n uint32) bool {
+	return rcv._tab.MutateUint32Slot(38, n)
+}
+
 func ObjectMetaStart(builder *flatbuffers.Builder) {
-	builder.StartObject(17)
+	builder.StartObject(18)
 }
 func ObjectMetaAddKey(builder *flatbuffers.Builder, key flatbuffers.UOffsetT) {
 	builder.PrependUOffsetTSlot(0, flatbuffers.UOffsetT(key), 0)
@@ -341,6 +353,9 @@ func ObjectMetaAddTags(builder *flatbuffers.Builder, tags flatbuffers.UOffsetT) 
 }
 func ObjectMetaStartTagsVector(builder *flatbuffers.Builder, numElems int) flatbuffers.UOffsetT {
 	return builder.StartVector(4, numElems, 4)
+}
+func ObjectMetaAddStripeBytes(builder *flatbuffers.Builder, stripeBytes uint32) {
+	builder.PrependUint32Slot(17, stripeBytes, 0)
 }
 func ObjectMetaEnd(builder *flatbuffers.Builder) flatbuffers.UOffsetT {
 	return builder.EndObject()